requirements.txt
Flask
Jinja2
MarkupSafe
Werkzeug
celery
certifi
chardet
configparser
gevent
gunicorn
idna
itsdangerous
lxml
matplotlib
mysql-connector
pillow
pip
pylint-fail-under
pytest
pytest-flask
pytest-xdist
httpx
setuptools
urllib3
wheel